Last refactor webpack config
